# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "Complete API for mechanic shop management with inventory tracking, JWT authentication, rate limiting, and caching"
  version: "1.0.0"
  contact:
    name: "Mechanic Shop API Support"
    email: "support@mechanicshop.com"
host: "be1-mechanic-shop-assignment.onrender.com"
schemes:
  - "https"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  Bearer:
    type: apiKey
    name: Authorization
    in: header
    description: |
      JWT Authorization header using the Bearer scheme.
      
      **Format:** `Bearer {jwt_token}`
      
      **Example:** `Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...`
      
      **How to obtain token:**
      1. Make a POST request to `/login` with valid credentials
      2. Copy the `access_token` from the response
      3. Add it to the Authorization header as `Bearer {access_token}`
      
      **Token contains:**
      - User identity and role information
      - Expiration time (typically 24 hours)
      - Permissions and access level
      
      **Usage:**
      - Include this header in requests to protected endpoints
      - Token will be validated on each request
      - Expired tokens will return 401 Unauthorized
  
  JWTAuth:
    type: apiKey
    name: Authorization
    in: header
    description: |
      JSON Web Token (JWT) authentication for API access.
      
      **Authentication Flow:**
      1. **Login:** POST credentials to `/login` endpoint
      2. **Token:** Receive JWT access token in response
      3. **Header:** Include token in Authorization header
      4. **Access:** Make authenticated requests to protected resources
      
      **Token Security:**
...
paths:
...
  /service-tickets/:
    get:
      tags:
        - "Service Tickets"
      summary: "Get all service tickets"
      responses:
        200:
          description: "Service tickets retrieved successfully"
    post:
      tags:
        - "Service Tickets"
      summary: "Create service ticket"
      parameters:
        - in: body
          name: ticket
          required: true
          schema:
            $ref: "#/definitions/ServiceTicketCreate"
      responses:
        201:
          description: "Service ticket created successfully"
        400:
          description: "Validation error"
  /service-tickets/{id}:
    get:
      tags:
        - "Service Tickets"
      summary: "Get service ticket by ID"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
      responses:
        200:
          description: "Service ticket retrieved successfully"
        404:
          description: "Service ticket not found"
  /service-tickets/{id}/inventory:
    post:
      tags:
        - "Service Tickets"
      summary: "Add inventory to ticket"
      description: "Add inventory parts to a service ticket"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: inventory_data
          required: true
          schema:
            type: object
            properties:
              inventory_ids:
                type: array
                items:
                  type: integer
                example: [1, 2, 3]
      responses:
        200:
          description: "Inventory added to ticket successfully"
        400:
          description: "Validation error"
        404:
          description: "Service ticket not found"
    delete:
      tags:
        - "Service Tickets"
      summary: "Remove inventory from ticket"
      description: "Remove inventory parts from a service ticket"
      parameters:
        - in: path
          name: id
          type: integer
          required: true
        - in: body
          name: inventory_data
          required: true
          schema:
            type: object
            properties:
              inventory_ids:
                type: array
                items:
                  type: integer
                example: [1, 2]
      responses:
        200:
          description: "Inventory removed from ticket successfully"
        400:
          description: "Validation error"
        404:
          description: "Service ticket not found"
  /service-tickets/{ticket_id}/assign-mechanic/{mechanic_id}:
    put:
      tags:
        - "Service Tickets"
      summary: "Assign mechanic to ticket"
      description: "Assign a mechanic to a service ticket (Rate Limited: 20 per minute)"
      parameters:
        - in: path
          name: ticket_id
          type: integer
          required: true
        - in: path
          name: mechanic_id
          type: integer
          required: true
      responses:
        200:
          description: "Mechanic assigned successfully"
        400:
          description: "Mechanic already assigned"
        404:
          description: "Service ticket or mechanic not found"
        429:
          description: "Rate limit exceeded"
  /service-tickets/{id}/edit:
    put:
      tags:
        - "Service Tickets"
      summary: "Bulk edit mechanics on a ticket"
      description: "Bulk add and remove mechanics from a service ticket."
      parameters:
        - in: path
          name: id
          type: integer
          required: true
          description: "The ID of the service ticket to edit."
        - in: body
          name: mechanic_ids
          required: true
          schema:
            type: object
            properties:
              add_ids:
                type: array
                items:
                  type: integer
                description: "List of mechanic IDs to add to the ticket."
                example: [1, 2]
              remove_ids:
                type: array
                items:
                  type: integer
                description: "List of mechanic IDs to remove from the ticket."
                example: [3]
      responses:
        200:
          description: "Mechanics updated successfully."
        207:
          description: "Partial success, some mechanics could not be added/removed."
        400:
          description: "Bad request or no changes made."
        404:
          description: "Service ticket not found."
definitions:
...
  ServiceTicketCreate:
    type: object
    required:
      - customer_id
      - description
      - service_date
    properties:
      customer_id:
        type: integer
        example: 1
      description:
        type: string
        example: "Oil change and brake inspection"
      service_date:
        type: string
        format: date
        example: "2024-12-24"
      mechanic_ids:
        type: array
        items:
          type: integer
        example: [1, 2]

  # Authentication-related definitions
...